Fixes and performance improvements for Dedupe Delta Compression Avoid using fingerprints in minhash computation and fix write amplification Modify min-heap to use 64bit values Improve bsdiff performance Fix pointer comparison in bsdiff Use 32bit offsets in bsdiff to reduce memory usage Improve Zero RLE Encoder performance Add more buffer overflow checks in Zero RLE Decoder

+++ utils/utils.h
@@ -57,7 +57,7 @@ extern "C" {
 #       endif
 #endif
 typedef unsigned long uintptr_t;
-typedef int64_t bsize_t;
+typedef int32_t bsize_t;
 
 #undef WORDS_BIGENDIAN
 #if BYTE_ORDER == BIG_ENDIAN
